# docs/swagger.yaml
definitions:
  router.Response:
    properties:
      data: {}
      msg:
        type: string
      status:
        type: integer
    type: object
info:
  contact:
    email: ChainBaby
    name: ChainBaby
  description: ChainBaby
  termsOfService: http://swagger.io/terms/
  title: ChainBaby
  version: "1.0"
paths:
  /api/approveCoin:
    get:
      description: 代币授权
      parameters:
      - description: 公链的ID 0代表以太 1代表币安 2代表波场
        in: query
        name: chain_id
        required: true
        type: integer
      - description: 代币地址
        in: query
        name: token_address
        required: true
        type: string
      - description: 私钥
        in: query
        name: private_key
        required: true
        type: string
      - description: 消费者地址
        in: query
        name: spender_address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            $ref: '#/definitions/router.Response'
      tags:
      - 代币授权
...
  /api/checkAllowance:
    get:
      description: 查询检查授权
      parameters:
      - description: 公链的ID 0代表以太 1代表币安 2代表波场
        in: query
        name: chain_id
        required: true
        type: integer
      - description: 代币地址
        in: query
        name: token_address
        required: true
        type: string
      - description: 来源地址
        in: query
        name: owner_address
        required: true
        type: string
      - description: 消费者地址
        in: query
        name: spender_address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            $ref: '#/definitions/router.Response'
      tags:
      - 查询Token 检查授权
...
  /api/getBaseAddress:
    get:
      description: 私钥转地址
      parameters:
      - description: 公链的ID 0代表以太 1代表币安
        in: query
        name: chain_id
        required: true
        type: integer
      - description: 私钥
        in: query
        name: private_key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            $ref: '#/definitions/router.Response'
      tags:
      - 私钥转地址
...
  /api/issueSimpleCoin:
    get:
      description: 发布普通代币
      parameters:
      - description: 公链的ID 0代表以太 1代表币安
        in: query
        name: chain_id
        required: true
        type: integer
      - description: 代币名称
        in: query
        name: name
        required: true
        type: string
      - description: 代币精度
        in: query
        name: decimal
        required: true
        type: integer
      - description: 代币符号
        in: query
        name: symbol
        required: true
        type: string
      - description: 总供应量
        in: query
        name: total_supply
        required: true
        type: integer
      - description: 私钥
        in: query
        name: private_key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            $ref: '#/definitions/router.Response'
      tags:
      - 发布普通代币
...
  /api/transferFromCoin:
    get:
      description: 授权转账代币
      parameters:
      - description: 公链的ID 0代表以太 1代表币安 2代表波场
        in: query
        name: chain_id
        required: true
        type: integer
      - description: 代币地址
        in: query
        name: token_address
        required: true
        type: string
      - description: 代币数量
        in: query
        name: amount
        required: true
        type: number
      - description: 来源地址
        in: query
        name: from_address
        required: true
        type: string
      - description: 目标地址
        in: query
        name: to_address
        required: true
        type: string
      - description: 消费地址私钥
        in: query
        name: private_key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            $ref: '#/definitions/router.Response'
      tags:
      - 授权转账代币
swagger: "2.0"